requirements.txt
django
djangorestframework
markdown
django-filter
requests
python-dotenv